requirements.txt
flask
flask-session
openai
python_dotenv
requests
beautifulsoup4
chromadb